text: Adds TextRecord
This patch adds a record type simply to split a text record into its
content and trailing entries sections.

In addition, the parsing routine of MobiFile is modified to utilise the
new class and extended to extract the book name from the offset
described by the MobiDocHeader. Finally, the IndexRecord[] is split
since only the first record contains a TagxHeader for use in decoding
all IDXT entries and that record usually only contains a dummy entry.